wip: Refactor websocket room create msg listener
